+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # UNRELEASED
 
+## Enhancements
+* Added BETA support for including `projects` relationship and `projects-count` attribute to policy_set on create by @hs26gill [#737](https://github.com/hashicorp/go-tfe/pull/737)
+
 # v1.30.0
 
 ## Enhancements